fix: 하단 탭바 scenedelegate에서 분리, add: stampservice.swift
